Create farmer
so in this commit, I create a function for Employee to create farmer and he can view. but this commit more create only were employee allows to create a farmer to register so farmer can add products. and also I even make it more professional look in the create view